Update endpoints of xmart databases for invasive species and translations.
The new endpoints correspond to the more mature tables that
are not TEST or TRAINING anymore.

The VECTORS_IR/TRANSLATIONS contains the same data as the
old TRAINING_EYESEETEA/MTM_TRADS, including the stories
for the Insecticide Resistance theme.

The VECTORS_IR/FACT_INVASIVE_SPECIES_LEGACY contains
the data from the view that presents the same data as
the old arcgis endpoint.