Fix: use maximum max for all charts
Summary of molecular marker study results over time	 - y-axis
Is it possible for the y-axis of all the charts to all be fixed at the same maximum within this selection? So for example, if the maximum number of studies in a year in a single country was 20, then all the y-axes should have the same maximum of around 20, even if only very few studies have been carried out.