Fix: data sources order in bars
Data source: ideally the first bar at the top of the chart is marked (a). It looks like the data sources are listed alphabetically, and hence the letters in the chart appear disorganised. Although in some charts it does not seemto be alphabetical, so I am unsure how the ordering is working. preference to begin sort by date (descending from most recent), then  insecticide type, as it is shown in the development screenshot. the  letter labels should as much as possible be in alphabetical order:  starting with a, b, c, etc. however if there are repeated citations then  the letter labels will also have to repeat. the order of appearance of citations listed under data source, should nonetheless be sorted alphabetically.