Fi: prevention synergist chart in side panel
These bars are grouped in pairs and the order should always be "No synergist", then "PBO". It was decided that we would just use the light grey and dark grey (using the same system as the current interface) and not the pink and purple in the bars. Lucia to confirm that this decision has not changed. Lucia would like a line of text added above the chart. This line of text would relate to the colour of the dot on the map. In this example here, it says "PBO partially restores susceptibility to Pyrethroids" The words in colour would change depending on the situation. Lucia to confirm the exact wording for the three options for the line of text are.