Fix: pfhrp2/3 gene deletions - Side panel
"If the sample is imported from a different country, this should be indicated above the chart. This information is in the existing MTM interface.

In column F, there is an a) before the data source which does not seem to refer to anything. If there is just one data source we can present it as in the design (column E) without a letter."