Fix: prevention dashboards filters
There's something going on with the filters: when you change the filters to "Only include bioassays with more than 150+ mosquitoes" and "Only include data provided by ministries of health", the number of studies increases. One would expect the number of studies to decrease because you're limiting the selection. https://share.vidyard.com/watch/X1rNCwtvEaVHMeNxchGA9A? The number of studies shown on the chart also increase for Carbamates when you remove the other insecticide classes from the list.